test: 테스트 케이스 위치 이관
